multi-agent/requirements.txt
# Core dependencies
langchain>=0.3.0
langgraph>=0.6.0,<1.0.0
langgraph-checkpoint-firestore>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.5

# Data and API
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
tenacity>=8.2.0
cachetools>=5.3.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
structlog>=23.1.0

# Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Storage
google-cloud-firestore>=2.16.0
google-cloud-secret-manager>=2.16.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
faker>=19.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.5.0